refactor(api): use typings from shared folder
