+++ functions/api/auth/reset-password.ts
@@ -1,20 +1,10 @@
-import { z } from "zod";
-import { User } from "../../models/user";
 import * as bcrypt from "bcryptjs";
+
+import { ResetPasswordRequest, SuccessResponse } from "@/shared/api/types";
+import { User } from "@/shared/models/user";
 import type { AppFunction } from "../../types/appcontext";
 import { HttpError } from "../../types/httperror";
 import { parseBody } from "../../utils/bodyparser";
-
-export const ResetPasswordRequest = z.object({
-  email: z.string().email(),
-  code: z.string(),
-  password: User.shape.password,
-});
-export type ResetPasswordRequest = z.infer<typeof ResetPasswordRequest>;
-
-export type ResetPasswordResponse = {
-  success: boolean;
-};
 
 export const onRequestPost: AppFunction = async ctx => {
   const req = await parseBody(ctx.request, ResetPasswordRequest);
@@ -41,6 +31,6 @@ export const onRequestPost: AppFunction = async ctx => {
   delete user.verificationCode;
   await ctx.env.users.put(user.email, JSON.stringify(user));
 
-  const resBody: ResetPasswordResponse = { success: true };
+  const resBody: SuccessResponse = { success: true };
   return Response.json(resBody);
 };
