+++ functions/api/users/me.ts
@@ -1,22 +1,12 @@
 import * as bcrypt from "bcryptjs";
-import { SanitizedUser, User } from "../../models/user";
+
+import { SanitizedUser, UpdatableUser } from "@/shared/models/user";
 import type { AppFunction } from "../../types/appcontext";
 import { ServerError } from "../../types/httperror";
 import { parseBody } from "../../utils/bodyparser";
 import { sendConfirmationEmail } from "../../utils/postmark";
 
 export const onRequestGet: AppFunction = ctx => Response.json(SanitizedUser.parse(ctx.data.user));
-
-const UpdatableUser = User.pick({
-  fullName: true,
-  preferredName: true,
-  email: true,
-  password: true,
-  currency: true,
-  interest_rate_type: true,
-  language: true,
-  timezone: true,
-}).partial();
 
 export const onRequestPatch: AppFunction = async ctx => {
   const patch = await parseBody(ctx.request, UpdatableUser);
