+++ functions/tsconfig.json
@@ -7,6 +7,10 @@
     "target": "ESNext",
     "module": "ESNext",
     "lib": ["ESNext"],
-    "types": ["@cloudflare/workers-types"]
+    "types": ["@cloudflare/workers-types"],
+    "baseUrl": "../",
+    "paths": {
+      "@/shared/*": ["shared/*"]
+    }
   }
 }
