+++ functions/utils/bodyparser.ts
@@ -1,4 +1,5 @@
 import { z, ZodObject, ZodRawShape } from "zod";
+
 import { ValidationError } from "../types/httperror";
 
 export const parseBody = async <R extends ZodRawShape, T extends ZodObject<R>>(
