+++ functions/utils/postmark.ts
@@ -1,7 +1,8 @@
-import type { AppContext } from "../types/appcontext";
 import { v4 as uuidv4 } from "uuid";
 import { z } from "zod";
-import type { User } from "../models/user";
+
+import type { User } from "@/shared/models/user";
+import type { AppContext } from "../types/appcontext";
 
 export type PostmarkHeader = {
   Name: string;
