+++ shared/api/types/index.ts
@@ -1,3 +1,2 @@
 export * from "./auth";
 export * from "./common";
-export * from "./users";
