--- shared/api/types/users/index.ts
@@ -1,1 +1,0 @@
-export * from "./me";
